split out developer panel code
