+++ selfdrive/ui/qt/offroad/developer_panel.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "selfdrive/ui/qt/offroad/settings.h"
+
+class DeveloperPanel : public ListWidget {
+  Q_OBJECT
+public:
+  explicit DeveloperPanel(SettingsWindow *parent);
+
+private:
+  ButtonControl *hellolBtn;
+};
